Split direct buildability policies
- Move backend-specific checks behind an internal buildability boundary

- Preserve direct readiness diagnostics behavior

- Tighten compiler metrics budgets around the split
