+++ scripts/compiler-metrics.mts
@@ -22,8 +22,10 @@ const fileBudgets = {
   "native/zero-c/src/ir.c": { maxLines: 3700, maxStrcmpCalls: 224 },
   "native/zero-c/src/row_syntax.c": { maxLines: 2150, maxStrcmpCalls: 11 },
   "native/zero-c/src/ast.c": { maxLines: 250, maxStrcmpCalls: 0 },
-  "native/zero-c/src/buildability.c": { maxLines: 610, maxStrcmpCalls: 20 },
+  "native/zero-c/src/buildability.c": { maxLines: 260, maxStrcmpCalls: 2 },
   "native/zero-c/src/buildability.h": { maxLines: 20, maxStrcmpCalls: 0 },
+  "native/zero-c/src/buildability_internal.h": { maxLines: 50, maxStrcmpCalls: 0 },
+  "native/zero-c/src/buildability_targets.c": { maxLines: 340, maxStrcmpCalls: 16 },
   "native/zero-c/src/call_resolve.c": { maxLines: 200, maxStrcmpCalls: 2 },
   "native/zero-c/src/call_resolve.h": { maxLines: 100, maxStrcmpCalls: 0 },
   "native/zero-c/src/coff_format.c": { maxLines: 370, maxStrcmpCalls: 0 },
